Refactor and modularize SVO navigation source files
Reorganized the SVONavigation source files by moving and renaming several implementation files into more granular and modular directories (e.g., Core, Common). Added new implementation files for node types and navigation data generators, and split large source files into smaller, focused components for SVO data, node types, and volume navigation data. This improves code maintainability and separation of concerns.
