+++ Source/SVONavigation/Private/Core/SVOVolumeNavigationData/SVOVolumeNavigationData.cpp
@@ -1,0 +1,7 @@
+﻿#include "SVOVolumeNavigationData.h"
+
+void FSVOVolumeNavigationData::Reset()
+{
+	VolumeBounds.Init();
+	SVOData.Reset();
+}
