+++ Source/SVONavigation/Public/SVOHelpers.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "SVONavigationTypes.h"
-
 #include <CoreMinimal.h>
 #include <GraphAStar.h>
 #include <ThirdParty/libmorton/morton.h>
